requirements.txt
async-timeout==4.0.2
certifi==2021.10.8
chardet==3.0.4
Deprecated==1.2.13
idna==2.10
newscatcherapi==0.5.0
packaging==21.3
pyparsing==3.0.7
PyYAML==6.0
redis==4.2.2
requests==2.24.0
urllib3==1.25.11
wrapt==1.14.0
